requirements.txt
Babel==2.6.0
boto3==1.9.62
botocore==1.12.62
dj-database-url==0.5.0
Django==2.1.3
django-bootstrap4==0.0.7
django-filter==2.0.0
django-phonenumber-field==2.1.0
django-storages==1.7.1
django-widget-tweaks==1.4.3
docutils==0.14
gunicorn==19.9.0
jmespath==0.9.3
phonenumbers==8.10.1
Pillow==5.3.0
psycopg2==2.7.6.1
python-dateutil==2.7.5
python-decouple==3.1
pytz==2018.7
s3transfer==0.1.13
six==1.11.0
urllib3==1.24.1
